+++ examples/src/test/java/org/bindgen/example/GenericsExampleTest.java
@@ -1,30 +1,33 @@
 package org.bindgen.example;
 
-import org.junit.Assert;
-import junit.framework.TestCase;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class GenericsExampleTest extends TestCase {
+import org.junit.jupiter.api.Test;
 
+public class GenericsExampleTest {
+
+	@Test
 	public void testField() {
 		GenericsExample<String> ge = new GenericsExample<String>();
 		GenericsExampleBinding<String> geb = new GenericsExampleBinding<String>(ge);
 
 		ge.bar = "1";
-		Assert.assertEquals("1", geb.bar().get());
+		assertEquals("1", geb.bar().get());
 
 		geb.bar().set("2");
-		Assert.assertEquals("2", ge.bar);
+		assertEquals("2", ge.bar);
 	}
 
+	@Test
 	public void testMethod() {
 		GenericsExample<String> ge = new GenericsExample<String>();
 		GenericsExampleBinding<String> geb = new GenericsExampleBinding<String>(ge);
 
 		ge.setFoo("1");
-		Assert.assertEquals("1", geb.foo().get());
+		assertEquals("1", geb.foo().get());
 
 		geb.foo().set("2");
-		Assert.assertEquals("2", ge.getFoo());
+		assertEquals("2", ge.getFoo());
 	}
 
 }