+++ examples/src/test/java/org/bindgen/example/inheritance/BaseExampleTest.java
@@ -1,34 +1,39 @@
 package org.bindgen.example.inheritance;
 
-import org.junit.Assert;
-import junit.framework.TestCase;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
-public class BaseExampleTest extends TestCase {
+import org.junit.jupiter.api.Test;
 
+public class BaseExampleTest {
+
+	@Test
 	public void testSubBindings() {
 		SubExample sub = new SubExample();
 		SubExampleBinding subb = new SubExampleBinding(sub);
 		subb.name().set("foo");
 		subb.subOnly().set("bar");
 
-		Assert.assertEquals("foo", sub.name);
-		Assert.assertEquals("bar", sub.subOnly);
+		assertEquals("foo", sub.name);
+		assertEquals("bar", sub.subOnly);
 
 		// because of the clash with the base class 'name', we get an extra 'nameField' that still points to 'SubExample.name'
 		subb.nameField().set("foo");
-		Assert.assertEquals("foo", sub.name);
+		assertEquals("foo", sub.name);
 
 		// 5 == base description, sub name, sub subOnly, hashCode and toString
 		// +1 currently for the parent name and child nameField
-		Assert.assertEquals(6, subb.getChildBindings().size());
+		assertEquals(6, subb.getChildBindings().size());
 	}
 
+	@Test
 	public void testSubBindingsWithRealSub() {
 		SubExampleBinding subb = new SubExampleBinding();
 		subb.set(new SubExample());
-		Assert.assertEquals(null, subb.subOnly().get());
+		assertNull(subb.subOnly().get());
 	}
 
+	@Test
 	public void testSubBindingsWithBase() {
 		// SubExampleBinding subb = new SubExampleBinding();
 		try {