+++ examples/src/test/java/org/bindgen/example/methods/MethodExample2Test.java
@@ -1,9 +1,12 @@
 package org.bindgen.example.methods;
 
-import junit.framework.TestCase;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class MethodExample2Test extends TestCase {
+import org.junit.jupiter.api.Test;
 
+public class MethodExample2Test {
+
+	@Test
 	public void testThreeNames() {
 		MethodExample2Binding b = new MethodExample2Binding(new MethodExample2());
 		assertEquals("1", b.name().get());