+++ examples/src/test/java/org/bindgen/example/methods/MethodWithGenericsExampleTest.java
@@ -1,22 +1,25 @@
 package org.bindgen.example.methods;
 
+import static org.junit.jupiter.api.Assertions.assertNotSame;
+import static org.junit.jupiter.api.Assertions.assertSame;
+
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Assert;
-import junit.framework.TestCase;
+import org.junit.jupiter.api.Test;
 
-public class MethodWithGenericsExampleTest extends TestCase {
+public class MethodWithGenericsExampleTest {
 
+	@Test
 	public void testReadWrite() {
 		MethodWithGenericsExample e = new MethodWithGenericsExample();
 		MethodWithGenericsExampleBinding b = new MethodWithGenericsExampleBinding(e);
 
 		List<String> originalList = e.getList();
-		Assert.assertSame(originalList, b.list().get());
+		assertSame(originalList, b.list().get());
 
 		b.list().set(new ArrayList<String>());
-		Assert.assertNotSame(originalList, b.list().get());
+		assertNotSame(originalList, b.list().get());
 	}
 
 }