+++ processor/src/test/java/org/bindgen/processor/AccessTest.java
@@ -1,8 +1,9 @@
 package org.bindgen.processor;
 
-import static org.junit.Assert.*;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.Test;
 
 public class AccessTest extends AbstractBindgenTestCase {