+++ processor/src/test/java/org/bindgen/processor/ConflictingGenericTypeArgumentTest.java
@@ -1,6 +1,6 @@
 package org.bindgen.processor;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  * @author Nándor Előd Fekete