+++ processor/src/test/java/org/bindgen/processor/GenericsMultipleBoundsTest.java
@@ -1,6 +1,6 @@
 package org.bindgen.processor;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class GenericsMultipleBoundsTest extends AbstractBindgenTestCase {