+++ processor/src/test/java/org/bindgen/processor/RootChangeTest.java
@@ -1,8 +1,10 @@
 package org.bindgen.processor;
 
-import static org.junit.Assert.*;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import org.junit.jupiter.api.Test;
 
 public class RootChangeTest extends AbstractBindgenTestCase {