+++ processor/src/test/java/org/bindgen/processor/util/ClassNameTest.java
@@ -1,20 +1,20 @@
 package org.bindgen.processor.util;
 
-import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.*;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class ClassNameTest {
 
 	@Test
 	public void stripRepeatedTypeVarsOfOuterClasses() {
-		assertThat(//
-			new ClassName("java.util.Map<K, V>.Entry<K, V>").get(),
-			is("java.util.Map.Entry<K, V>"));
-		assertThat(//
-			new ClassName("java.util.Foo<K extends java.util.Bar<K>>.Entry<K extends java.util.Bar<K>>").get(),
-			is("java.util.Foo.Entry<K extends java.util.Bar<K>>"));
+		assertEquals(
+        "java.util.Map.Entry<K, V>",
+        new ClassName("java.util.Map<K, V>.Entry<K, V>").get());
+    assertEquals(
+        "java.util.Foo.Entry<K extends java.util.Bar<K>>",
+        new ClassName("java.util.Foo<K extends java.util.Bar<K>>.Entry<K extends java.util.Bar<K>>").get());
 	}
 
 }